[move] Move core parsing tests into separate file
