+++ external-crates/move/crates/move-core-types/src/unit_tests/mod.rs
@@ -5,5 +5,6 @@
 mod extractor_test;
 mod identifier_test;
 mod language_storage_test;
+mod parsing_test;
 mod value_test;
 mod visitor_test;
